% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_total_orcamento.R
\name{check_total_orcamento_fiscal}
\alias{check_total_orcamento_fiscal}
\title{Total do orçamento fiscal e investimento SIGPLAN vs SISOR}
\usage{
check_total_orcamento_fiscal(
  base_qdd_fiscal,
  acoes_planejamento,
  verbose = NULL
)
}
\description{
Verificar se o valor total do orçamento (orçamento fiscal e orçamento de
investimento das empresas controladas) coincide com a projeção do PPAG
para o ano seguinte
}
